Move DataSourcePreparer
